Replace agent dropdown with Multiselect, update mobile CSS
