+++ src/scss/_reset.scss
@@ -12,8 +12,6 @@ i {
 
 .page-container {
   position: relative;
-  background-color: rgb(255, 251, 240);
-  // max-width: 80rem;
-  // margin: 1rem auto 5rem auto;
+  background-color: #fffbf0;
   margin: 0 auto;
 }
