+++ src/scss/_search-form.scss
@@ -36,12 +36,19 @@
   width: 90%;
   max-width: 75rem;
   min-height: 11.5rem;
-  background-color: rgba(255, 255, 255, 0.97);
+  background-color: rgba(255, 255, 255, 0.96);
   border-radius: 0.6rem;
   box-shadow: 0 0 0 0.5rem rgba(255, 255, 255, 0.25);
   margin: 3rem auto 0 auto;
-  padding: 2.5rem;
+  padding: 2.5rem 6rem;
   z-index: 55;
+}
+
+.search-form-row {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
 }
 
 .search-label {
@@ -170,15 +177,15 @@
   z-index: 60;
 
   &:hover {
-    background-color: rgb(215, 235, 252);
-    border-color: rgb(215, 235, 252);
+    background-color: rgb(239, 246, 252);
+    border-color: rgb(239, 246, 252);
   }
 }
 
 .search-dropdown-option-selected {
   position: relative;
   font-size: 1.1rem;
-  background-color: rgb(172, 215, 250);
+  background-color: rgb(196, 228, 255);
   width: 100%;
   text-align: left; 
   padding: 0.2rem 1.25rem 0.2rem 0.5rem;
